Move theme validation to utilk method
